+++ LuckyoneUI/Modules/Chat/Chat.lua
@@ -11,20 +11,21 @@ local SetCVar = C_CVar.SetCVar
 
 -- Global environment
 local _G = _G
+local ChatFrameUtil = _G.ChatFrameUtil
 
 -- Blizzard functions
 local ChatFrame_AddMessageGroup = ChatFrame_AddMessageGroup
 local ChatFrame_RemoveAllMessageGroups = ChatFrame_RemoveAllMessageGroups
 local ChatFrame_RemoveMessageGroup = ChatFrame_RemoveMessageGroup
-local FCF_DockFrame = FCF_DockFrame
-local FCF_OpenNewWindow = FCF_OpenNewWindow
-local FCF_ResetChatWindow = FCF_ResetChatWindow
-local FCF_ResetChatWindows = FCF_ResetChatWindows
-local FCF_SavePositionAndDimensions = FCF_SavePositionAndDimensions
-local FCF_SetChatWindowFontSize = FCF_SetChatWindowFontSize
-local FCF_SetWindowName = FCF_SetWindowName
-local FCF_StopDragging = FCF_StopDragging
-local FCFDock_SelectWindow = FCFDock_SelectWindow
+local FCF_DockFrame = ChatFrameUtil and ChatFrameUtil.DockFrame or FCF_DockFrame
+local FCF_OpenNewWindow = ChatFrameUtil and ChatFrameUtil.OpenNewWindow or FCF_OpenNewWindow
+local FCF_ResetChatWindow = ChatFrameUtil and ChatFrameUtil.ResetChatWindow or FCF_ResetChatWindow
+local FCF_ResetChatWindows = ChatFrameUtil and ChatFrameUtil.ResetChatWindows or FCF_ResetChatWindows
+local FCF_SavePositionAndDimensions = ChatFrameUtil and ChatFrameUtil.SavePositionAndDimensions or FCF_SavePositionAndDimensions
+local FCF_SetChatWindowFontSize = ChatFrameUtil and ChatFrameUtil.SetChatWindowFontSize or FCF_SetChatWindowFontSize
+local FCF_SetWindowName = ChatFrameUtil and ChatFrameUtil.SetWindowName or FCF_SetWindowName
+local FCF_StopDragging = ChatFrameUtil and ChatFrameUtil.StopDragging or FCF_StopDragging
+local FCFDock_SelectWindow = ChatFrameUtil and ChatFrameUtil.Dock_SelectWindow or FCFDock_SelectWindow
 local VoiceTranscriptionFrame_UpdateEditBox = VoiceTranscriptionFrame_UpdateEditBox
 local VoiceTranscriptionFrame_UpdateVisibility = VoiceTranscriptionFrame_UpdateVisibility
 local VoiceTranscriptionFrame_UpdateVoiceTab = VoiceTranscriptionFrame_UpdateVoiceTab
@@ -36,6 +37,31 @@ local VOICE = VOICE
 local WHISPER_TAB = { 'WHISPER', 'BN_WHISPER', 'IGNORED' }
 local GUILD_TAB = { 'GUILD', 'GUILD_ACHIEVEMENT', 'OFFICER' }
 local PARTY_TAB = { 'PARTY', 'PARTY_LEADER', 'RAID', 'RAID_LEADER', 'RAID_WARNING', 'INSTANCE_CHAT', 'INSTANCE_CHAT_LEADER', 'SYSTEM' }
+
+-- Compatibility for Mists & Era
+local function AddMessageGroup(frame, group)
+	if frame.AddMessageGroup then
+		frame:AddMessageGroup(group)
+	else
+		ChatFrame_AddMessageGroup(frame, group)
+	end
+end
+
+local function RemoveAllMessageGroups(frame)
+	if frame.RemoveAllMessageGroups then
+		frame:RemoveAllMessageGroups()
+	else
+		ChatFrame_RemoveAllMessageGroups(frame)
+	end
+end
+
+local function RemoveMessageGroup(frame, group)
+	if frame.RemoveMessageGroup then
+		frame:RemoveMessageGroup(group)
+	else
+		ChatFrame_RemoveMessageGroup(frame, group)
+	end
+end
 
 -- Chat setup for tabs, windows and channels
 function Private:Setup_Chat(installer)
@@ -83,7 +109,7 @@ function Private:Setup_Chat(installer)
 
 		-- Tabs
 		if id == 1 then
-			FCF_SetWindowName(frame, 'General')
+			FCF_SetWindowName(frame, 'Main')
 			if Private.ElvUI then
 				frame:ClearAllPoints()
 				frame:Point('BOTTOMLEFT', _G.LeftChatToggleButton, 'TOPLEFT', 1, 3)
@@ -107,27 +133,27 @@ function Private:Setup_Chat(installer)
 	end
 
 	-- Whisper tab
-	ChatFrame_RemoveAllMessageGroups(_G.ChatFrame4)
+	RemoveAllMessageGroups(_G.ChatFrame4)
 	for _, v in ipairs(WHISPER_TAB) do
-		ChatFrame_AddMessageGroup(_G.ChatFrame4, v)
+		AddMessageGroup(_G.ChatFrame4, v)
 	end
 
 	-- Guild tab
-	ChatFrame_RemoveAllMessageGroups(_G.ChatFrame5)
+	RemoveAllMessageGroups(_G.ChatFrame5)
 	for _, v in ipairs(GUILD_TAB) do
-		ChatFrame_AddMessageGroup(_G.ChatFrame5, v)
+		AddMessageGroup(_G.ChatFrame5, v)
 	end
 
 	-- Party tab
-	ChatFrame_RemoveAllMessageGroups(_G.ChatFrame6)
+	RemoveAllMessageGroups(_G.ChatFrame6)
 	for _, v in ipairs(PARTY_TAB) do
-		ChatFrame_AddMessageGroup(_G.ChatFrame6, v)
+		AddMessageGroup(_G.ChatFrame6, v)
 	end
 
 	-- Remove whispers from main tab
-	ChatFrame_RemoveMessageGroup(_G.ChatFrame1, 'IGNORED')
-	ChatFrame_RemoveMessageGroup(_G.ChatFrame1, 'WHISPER')
-	ChatFrame_RemoveMessageGroup(_G.ChatFrame1, 'BN_WHISPER')
+	RemoveMessageGroup(_G.ChatFrame1, 'IGNORED')
+	RemoveMessageGroup(_G.ChatFrame1, 'WHISPER')
+	RemoveMessageGroup(_G.ChatFrame1, 'BN_WHISPER')
 
 	-- Select the main tab
 	FCFDock_SelectWindow(_G.GENERAL_CHAT_DOCK, _G.ChatFrame1)